requirements.txt
Django~=3.2.23
Pillow~=10.1.0
djangorestframework~=3.14.0
mysqlclient ~= 2.2.4